# gateway/openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Platform API
  description: |
    Comprehensive Healthcare Platform API for managing patients, providers, 
    appointments, claims, payments, and more.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the 
    `Authorization` header as `Bearer <token>`.
    
    ## Rate Limits
    - Anonymous endpoints: 30 requests/minute
    - Authenticated endpoints: 100 requests/minute
    - Admin endpoints: 30 requests/minute
    
    ## Versioning
    API version is specified in the URL path (e.g., `/api/v1/...`) or 
    via the `X-API-Version` header.
  version: 1.0.0
  contact:
    name: Healthcare Platform Support
    email: support@healthcare.com
    url: https://healthcare.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.healthcare.com/v1
    description: Production server
  - url: https://staging-api.healthcare.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: User already exists
  /auth/login:
    post:
      tags: [Auth]
      summary: User login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/logout:
    post:
      tags: [Auth]
      summary: User logout
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful

  # ==================== Providers ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for external integrations
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [patient, provider, admin]
        phone:
          type: string
        dateOfBirth:
          type: string
          format: date
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
...